requirements.txt
django>=4.2,<5
asgiref>=3.6.0,<4
sqlparse>=0.5.0
django-allauth
django-unfold
gunicorn
python-dotenv
PyJWT>=2.7
cryptography>=41
setuptools
requests
